Refactor material-ui custom theme #6
